Migrate out built-in charts logic
